back/requirements.txt
Flask==2.2.3
python_nmap==0.7.1
Werkzeug<3.0.0
flask_sqlalchemy
Flask-Migrate
flask-cors
reportlab